requirements.txt
certifi==2024.8.30
cffi==1.17.1
cryptography==44.0.0
decorator==5.1.1
future==1.0.0
pycparser==2.22
python-telegram-bot==12.7
tornado==6.4.2
urllib3==1.26.20
emoji==2.14.0
